% Generated by roxygen2 (4.0.1): do not edit by hand
\name{convertInputColors}
\alias{convertInputColors}
\title{The R function for converting the data frame with colors to chromDraw color data format.}
\usage{
convertInputColors(colors);
}
\arguments{
\item{colors}{data frame definition of coloros, containing color name and RGB of color values.}
}
\value{
return file with colors in working directory.
}
\description{
The R function for converting the data frame with colors to chromDraw color data format. Output of this function is file "colors.txt" in the working directory.
}
\examples{
#colors generating
name <- c("yellow", "red", "blue", "violet", "orange", "green", "light_blue", "pink");
r <- c(255, 255, 0, 255, 247, 0, 0, 230);
g <- c(255, 0, 0, 0 ,148, 255, 255, 170);
b <- c(0, 0, 255, 255, 29, 0, 255, 160);
inputColors <- data.frame(name,r,g,b);

#run the function for generate chromDraw color file.
convertInputColors(inputColors);
}


