% Generated by roxygen2 (4.0.1): do not edit by hand
\name{main_chromDraw}
\alias{main_chromDraw}
\title{Main chromDraw function. The main function that controls all other C++ functions.}
\usage{
main_chromDraw(argc, argv)
}
\arguments{
\item{argc}{Count of input parameters.}

\item{argv}{Vector with input parameters strings.}
}
\value{
Return exit sate.
}
\description{
Main chromDrawS function. The main function that controls all other C++ functions.
}
\examples{
OUTPUTPATH = file.path(getwd());
INPUTPATH = system.file('extdata','Ack_and_Stenopetalum_nutans.txt', package ='chromDraw')
COLORPATH = system.file('extdata','default_colors.txt', package ='chromDraw')
chromDraw(argc=7, argv=c("chromDraw", "-c", COLORPATH, "-d", INPUTPATH, "-o", 
OUTPUTPATH));
}
