% Generated by roxygen2 (4.0.1): do not edit by hand
\name{chromDraw}
\alias{chromDraw}
\title{Main chromDraw function. Thsic R function call main C++ function.}
\usage{
chromDraw(argc, argv)
}
\arguments{
\item{argc}{count of input parameters.}

\item{argv}{vector with input parameters strings.}
}
\value{
return exit sate.
}
\description{
Main chromDraw function. Thsic R function call main C++ function.
}
\examples{
OUTPUTPATH = file.path(getwd());
INPUTPATH = system.file('extdata','Ack_and_Stenopetalum_nutans.txt', package ='chromDraw')
COLORPATH = system.file('extdata','default_colors.txt', package ='chromDraw')
chromDraw(argc=7, argv=c("chromDraw", "-c", COLORPATH, "-d", INPUTPATH, "-o", 
OUTPUTPATH));
}

