% Generated by roxygen2 (4.0.1): do not edit by hand
\name{main_chromDraw}
\alias{main_chromDraw}
\title{The main chromDraw function, that controls all other C++ functions.}
\usage{
main_chromDraw(argc, argv)
}
\arguments{
\item{argc}{Count of the input parameters.}

\item{argv}{Vector with the input parameters strings.}
}
\value{
Return exit state.
}
\description{
The main chromDraw function, that controls all other C++ functions.
}
\examples{
OUTPUTPATH = file.path(getwd());
INPUTPATH = system.file('extdata','Ack_and_Stenopetalum_nutans.txt', package ='chromDraw')
COLORPATH = system.file('extdata','default_colors.txt', package ='chromDraw')
chromDraw(argc=7, argv=c("chromDraw", "-c", COLORPATH, "-d", INPUTPATH, "-o", 
OUTPUTPATH));
}
